Inventory so far [Needs fixing]
Players can:
- Pick up items off the ground
- Double click on an item to equip it (BROKEN*)
- Devs can also create more items

Also added the keycard door type, which is WIP and just functions as a normal button for now.

Broken *
Anyway, for reasons I have yet to find out whenever the player double clicks an item to equip it, it always thinks that the item you equipped is the first item that the player picked up. So if  you picked up a keycard and then a cup, and then equipped the cup, the script will think you tried to equip the keycard and do that instead. This needs some urgent fixing. I was going to wait to make a GitHub push until the inventory was done but I don't want to destroy even the working stuff further.