Bringing this repo up to date
It's been a bit since I updated the repo as I wanted to finish the intro first. What I hoped to achieve for 0.0.2 is essentially done but there's a couple of problems with it right now, more info on the SCP:CBR Discord Server.

This push mainly features engine updates and new stuff for the intro.